Fix owner type

+++ distribution_strategy.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <algorithm>
+#include <functional>
 #include <stdexcept>
 #include <utility>
 #include <vector>
@@ -9,7 +10,7 @@ class DistributionStrategy {
 public:
   virtual ~DistributionStrategy() = default;
 
-  virtual size_t owner(size_t global_index) const = 0;
+  virtual int owner(size_t global_index) const = 0;
 
   virtual size_t local_size(int rank) const = 0;
 
@@ -25,7 +26,7 @@ public:
   BlockDistribution(std::vector<size_t> distribution) : _distribution(std::move(distribution)) {}
 
   // Worker with rank k stores the indices between _distribution[k] (inclusive) and _distribution[k + 1] (exclusive)
-  size_t owner(size_t global_index) const override {
+  int owner(size_t global_index) const override {
     auto it = std::upper_bound(_distribution.begin(), _distribution.end(), global_index);
     return std::distance(_distribution.begin(), it) - 1;
   }
@@ -46,7 +47,7 @@ class RoundRobinDistribution : public DistributionStrategy {
 public:
   RoundRobinDistribution(size_t total_elements, int num_ranks) : _total_elements(total_elements), _num_ranks(num_ranks) {}
 
-  size_t owner(size_t global_index) const override { return global_index % _num_ranks; }
+  int owner(size_t global_index) const override { return global_index % _num_ranks; }
 
   size_t local_size(int rank) const override {
     size_t local_elements = _total_elements / _num_ranks;
